local rank

+++ lib/config.py
@@ -94,20 +94,21 @@ check_dir(cfg.save_path)
 
 cfg.save_path = join(cfg.save_path, args.name)
 
-if not os.path.exists(cfg.save_path):
-    os.mkdir(cfg.save_path)
-elif not cfg.test:
-    key = input('Delete Existing Directory [y/n]: ')
-    if key == 'n':
-        if cfg.auto_resume:
-            pass
-        else:
-            raise ValueError("Save directory already exists")
-    elif key == 'y':
-        shutil.rmtree(cfg.save_path)
+if cfg.ddp.local_rank == 0:
+    if not os.path.exists(cfg.save_path):
         os.mkdir(cfg.save_path)
-    else:
-        raise ValueError("Input Not Supported!")
+    elif not cfg.test:
+        key = input('Delete Existing Directory [y/n]: ')
+        if key == 'n':
+            if cfg.auto_resume:
+                pass
+            else:
+                raise ValueError("Save directory already exists")
+        elif key == 'y':
+            shutil.rmtree(cfg.save_path)
+            os.mkdir(cfg.save_path)
+        else:
+            raise ValueError("Input Not Supported!")
 
 cfg.ckp_dir = check_dir(join(cfg.save_path, 'checkpoints'))
 cfg.log_dir = check_dir(join(cfg.save_path, 'runs'))
